add open_shell, transport_id
