+++ adbutils/_proto.py
@@ -6,7 +6,7 @@ from __future__ import annotations
 
 __all__ = [
     "Network", "BrightnessMode", "DeviceEvent", "ForwardItem", "ReverseItem", "FileInfo",
-    "WindowSize", "RunningAppInfo", "ShellReturn", "AdbDeviceInfo", "AppInfo", "BatteryInfo"
+    "WindowSize", "RunningAppInfo", "ShellReturn", "AdbDeviceInfo", "AppInfo", "BatteryInfo", "RunningAppInfo"
 ]
 
 import enum
@@ -127,6 +127,11 @@ class AdbDeviceInfo:
     serial: str
     state: str
     tags: dict[str, str] = field(default_factory=dict)
+    
+    @property
+    def transport_id(self) -> Optional[int]:
+        """Get the transport ID from tags."""
+        return int(self.tags.get("transport_id", 0)) if "transport_id" in self.tags else None
 
 
 StrOrPathLike = Union[str, pathlib.Path]
