+++ e2etests/test_device_shell_extend.py
@@ -9,57 +9,13 @@ import io
 import pathlib
 import re
 import time
-import filecmp
+
 import uuid
 
 import pytest
 
-import adbutils
 from adbutils import AdbDevice, Network, BrightnessMode
 from adbutils.errors import AdbSyncError
-
-
-def test_shell(device: AdbDevice):
-    for text in ("foo", "?", "&", "but    123"):
-        output = device.shell(['echo', '-n', text])
-        assert output == text
-
-    output = device.shell("pwd", rstrip=False)
-    assert output in ["/\n", "/\r\n"]
-
-
-def test_shell_without_encoding(device: AdbDevice):
-    output = device.shell("echo -n hello", encoding=None)
-    assert output == b"hello"
-
-    ret = device.shell2("echo -n hello", encoding=None)
-    assert ret.output == b"hello"
-
-
-def test_shell_stream(device: AdbDevice):
-    c = device.shell(["echo", "-n", "hello world"], stream=True)
-    output = c.read_until_close()
-    assert output == "hello world"
-
-
-def test_adb_shell_raise_timeout(device: AdbDevice):
-    with pytest.raises(adbutils.AdbTimeout):
-        device.shell("sleep 10", timeout=1.0)
-
-
-def test_shell2(device: AdbDevice):
-    cmd = "echo -n 'hello'; false"
-    res = device.shell2(cmd)
-    assert res.output == "hello"
-    assert res.returncode == 1
-    assert res.command == cmd
-
-
-def test_get_xxx(device: AdbDevice):
-    assert device.get_serialno()
-    assert device.get_state() == "device"
-    # adb connect device devpath is "unknown"
-    # assert device.get_devpath().startswith("usb:")
 
 
 def test_battery(device: AdbDevice):
@@ -202,11 +158,6 @@ def test_screenshot(device: AdbDevice):
     assert im.mode == "RGB"
 
 
-def test_framebuffer(device: AdbDevice):
-    im = device.framebuffer()
-    assert im.size
-
-
 def test_app_info(device: AdbDevice):
     pinfo = device.app_current()
     app_info = device.app_info(pinfo.package)
@@ -259,61 +210,3 @@ def test_logcat(device: AdbDevice, tmp_path: pathlib.Path):
     assert logcat_path.exists()
     assert re.compile(r"I/TAG.*hello").search(logcat_path.read_text(encoding="utf-8"))
 
-
-# todo: make independent of already present stuff on the phone
-def test_pull_push_dirs(
-        device: AdbDevice,
-        device_tmp_dir_path: str,
-        local_src_in_dir: pathlib.Path,
-        tmp_path: pathlib.Path,
-):
-    def are_dir_trees_equal(dir1, dir2):
-        """
-        Compare two directories recursively. Files in each directory are
-        assumed to be equal if their names and contents are equal.
-
-        NB: retreived from: https://stackoverflow.com/a/6681395
-
-        @param dir1: First directory path
-        @param dir2: Second directory path
-
-        @return: True if the directory trees are the same and 
-            there were no errors while accessing the directories or files, 
-            False otherwise.
-        """
-
-        dirs_cmp = filecmp.dircmp(dir1, dir2)
-        if len(dirs_cmp.left_only) > 0 or len(dirs_cmp.right_only) > 0 or \
-                len(dirs_cmp.funny_files) > 0:
-            return False
-        (_, mismatch, errors) = filecmp.cmpfiles(
-            dir1, dir2, dirs_cmp.common_files, shallow=False)
-        if len(mismatch) > 0 or len(errors) > 0:
-            return False
-        for common_dir in dirs_cmp.common_dirs:
-            new_dir1 = os.path.join(dir1, common_dir)
-            new_dir2 = os.path.join(dir2, common_dir)
-            if not are_dir_trees_equal(new_dir1, new_dir2):
-                return False
-        return True
-
-    local_src_out_dir1 = tmp_path / 'dir1'
-    local_src_out_dir2 = tmp_path / 'dir2'
-
-    # TODO: push src support dir
-    # device.push(local_src_in_dir, device_tmp_dir_path)
-    device.adb_output("push", str(local_src_in_dir), device_tmp_dir_path)
-
-    device.sync.pull_dir(device_tmp_dir_path, local_src_out_dir1)
-
-    assert local_src_out_dir1.exists()
-    assert local_src_out_dir1.is_dir()
-
-    are_dir_trees_equal(local_src_in_dir, local_src_out_dir1)
-
-    device.sync.pull(device_tmp_dir_path, local_src_out_dir2)
-
-    assert local_src_out_dir2.exists()
-    assert local_src_out_dir2.is_dir()
-
-    are_dir_trees_equal(local_src_in_dir, local_src_out_dir2)
